move common code
